ResourceNotification
New Resource Notification added.
The Resource Notification will update based on the user's  resource income.
Positive indicator for a positive income of that resource, and negative indicator for a negative income.

Neutral will need some work to indicate when a user is "Around 0 income" plus or minus a specific amount.